refactor(hooks): extract popup close logic into usePopupClose hook
