+++ src/components/Skills/index.js
@@ -2,6 +2,7 @@ import React, { useEffect, useRef, useState } from "react";
 import AnimatedLetters from "../AnimatedLetters";
 import "./index.scss";
 import { Helmet } from "react-helmet-async";
+import usePopupClose from "../../hooks/usePopupClose";
 import skillCategories from "../../data/skills";
 import {
   faClose,
@@ -47,42 +48,15 @@ const Skills = () => {
     getSkill();
   }, []);
 
-  useEffect(() => {
-    // If popup is opened and clicked outside the popup then close the popup
-    function handleClickOutside(event) {
-      if (popupRef.current && !popupRef.current.contains(event.target)) {
-        closePopup();
-      }
-    }
-
-    // If popup is opened and Esc key is pressed then close the popup
-    function handleEscKey(event) {
-      if (event.key === 'Escape' && showPopup) {
-        closePopup();
-      }
-    }
-
-    // Only add the event listener when the popup is shown
-    if (showPopup) {
-      document.addEventListener('mousedown', handleClickOutside);
-      document.addEventListener('keydown', handleEscKey);
-    }
-
-    // Clean up the event listener
-    return () => {
-      document.removeEventListener('mousedown', handleClickOutside);
-      document.removeEventListener('keydown', handleEscKey);
-    };
-  }, [showPopup]);
-
-  useScrollLock(showPopup);
-
   // Function to close popup and clean URL
   const closePopup = () => {
     setShowPopup(false);
     // Remove query parameters from URL
     // removeQueryParamFromURL();
   };
+
+  usePopupClose(popupRef, showPopup, closePopup);
+  useScrollLock(showPopup);
 
   const getSkill = async () => {
     setSkill(skillCategories.map((doc) => doc));
